test: разбил тесты по отдельным файлам для каждого класса
